show sale order

+++ Data/Purchase.cs
@@ -25,7 +25,7 @@ namespace Dukan.Data
         public int SupplierId { get; set; }
         public Nullable<System.DateTime> Date { get; set; }
         public int Number { get; set; }
-        public Nullable<double> Discount { get; set; }
+        public Nullable<decimal> Discount { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<PurchaseDetail> PurchaseDetails { get; set; }
