+++ Dukan.Core/Models/Sale/AddSaleDetailGridModel.cs
@@ -11,21 +11,21 @@ namespace Dukan.Core.Models.Sale
         public string Name { get; set; }
 
         [DisplayName("السعر")]
-        public double ProductSell { get; set; }
+        public decimal ProductSell { get; set; }
 
         [Browsable(false)]
-        public double ProductBuy { get; set; }
+        public decimal ProductBuy { get; set; }
 
         [DisplayName("الكمية")]
         public int? Qte { get; set; }
 
         [DisplayName("الإجمالي")]
-        public double Total { get; set; }
+        public decimal Total { get; set; }
 
         [DisplayName("الخصم")]
-        public double Discount { get; set; }
+        public decimal Discount { get; set; }
 
         [DisplayName("الإجمالي بعد الخصم")]
-        public double TotalAfterDiscount { get; set; }
+        public decimal TotalAfterDiscount { get; set; }
     }
 }
