+++ Dukan.Core/Models/Sale/SaleDetailReportModel.cs
@@ -4,11 +4,11 @@
     {
         public int Id { get; set; }
         public string ProductName { get; set; }
-        public double ProductSell { get; set; }
+        public decimal ProductSell { get; set; }
         public int Qte { get; set; }
-        public double Total { get; set; }
-        public double Discount { get; set; }
-        public double TotalAfterDiscount { get; set; }
+        public decimal Total { get; set; }
+        public decimal Discount { get; set; }
+        public decimal TotalAfterDiscount { get; set; }
         public bool IsReturned { get; set; }
     }
 }
